Break out catalog commands
